requirements.txt
requests
docopt
sendgrid
agrc-supervisor
